# add FZ16 to fpcr_fz_mask for aarch64