Fixes #40151 by also setting the FZ16 mask

@milankl can you test on your A64fx machine?