@milankl 2021-03-27
Works like a charm on a64fx