@milankl 2021-03-27
And last check, this PR also addresses the performance issues with subnormals and yields a decent 4x speed-up compared to Float64 without and with subnormals, whereas the latter previously slowed down things by 5x
```julia
julia> benchmark_float16_subnormals(100_000)
Without subnormals
  Float64:   88.040 μs (0 allocations: 0 bytes)
  Float32:   44.390 μs (0 allocations: 0 bytes)
  Float16:   22.351 μs (0 allocations: 0 bytes)
With 1% subnormals
  Float64:   88.240 μs (0 allocations: 0 bytes)
  Float32:   44.360 μs (0 allocations: 0 bytes)
  Float16:   22.400 μs (0 allocations: 0 bytes)
```
this just benchmarks adding two vectors{Float64/32/16} of length 100_000 element-wise.