@vchuravy 2021-03-27
> Have you checked if it is defined to write to reserved bits on chips without the fp16 feature? (I'll assume you are writing the correct bits)

The resource I based this off was https://developer.arm.com/documentation/100403/0201/register-descriptions/advanced-simd-and-floating-point-registers/aarch64-register-descriptions/fpcr--floating-point-control-register

If we want to be safe I can guard this behind a check for FP16 support (i.e. ID_AA64PFR0_EL1.)