@yuyichao 2021-03-27
> If we want to be safe I can guard this behind a check for FP16 support (i.e. ID_AA64PFR0_EL1.)

I couldn't find anything that says that the reserved bit is safe to access so I think we should safeguard this. Initializing a variable holding this flag is probably OK.
The system register isn't accessible from EL0 (as the name suggests), it's the `fullfp16` feature.