@vchuravy 2021-03-30
@milankl would be good to check whether this still works  on A64FX?