@milankl 2021-03-31
@vchuravy Yes, still works
```julia
julia> [floatmin(T)/2 for T in [Float64,Float32,Float16]]
3-element Vector{AbstractFloat}:
 1.1125369292536007e-308
 5.877472f-39
  Float16(3.05e-5)

julia> set_zero_subnormals(true)
true

julia> [floatmin(T)/2 for T in [Float64,Float32,Float16]]
3-element Vector{AbstractFloat}:
 0.0
 0.0f0
  Float16(0.0)
```